phase-e(port): src/jsc/web_worker.rs

+++ src/jsc/web_worker.rs
@@ -1161,7 +1161,7 @@ impl WebWorker {
                 // SAFETY: vm_ptr valid; sole owner on this thread.
                 rare.close_all_socket_groups(unsafe { &*vm_ptr });
             }
-            exit_code = vm.exit_handler.exit_code;
+            exit_code = i32::from(vm.exit_handler.exit_code);
             global_object = vm.global;
         }
